Move download audio option, clean up menus
