+++ apple/OmnivoreKit/Sources/App/Views/WebReader/WebReaderViewModel.swift
@@ -1,6 +1,7 @@
 import Models
 import Services
 import SwiftUI
+import Views
 import WebKit
 
 struct SafariWebLink: Identifiable {
@@ -20,6 +21,14 @@ struct SafariWebLink: Identifiable {
       return true
     }
     return false
+  }
+
+  func downloadAudio(audioController: AudioController, item: LinkedItem) {
+    Snackbar.show(message: "Downloading Offline Audio")
+    Task {
+      let downloaded = await audioController.downloadForOffline(itemID: item.unwrappedID)
+      Snackbar.show(message: downloaded ? "Audio file downloaded" : "Error downloading audio")
+    }
   }
 
   func loadContent(dataService: DataService, username: String, itemID: String, retryCount: Int = 0) async {
